Inline restaurant search panel
